update search and delete user

+++ src/main/java/com/humble/usercenter/controller/UserController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static com.humble.usercenter.constant.UserConstant.ADMIN_ROLE;
 import static com.humble.usercenter.constant.UserConstant.USER_LOGIN_STATE;
@@ -63,24 +64,38 @@ public class UserController {
     @GetMapping("/search")
     public List<User> searchUsers(String username, HttpServletRequest request) { // get请求 数据通过url传递 不包含请求体不需要@RequestBody注解
         // 鉴权 仅管理员可查询
-        Object userObj = request.getSession().getAttribute(USER_LOGIN_STATE);
-        User user = (User) userObj;
-        if (user == null || user.getRole() != ADMIN_ROLE) { // user == null 防止空指针异常 NullPointerException
+        if (!isAdmin(request)) {
             return new ArrayList<>();
         }
-
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
         if (StringUtils.isNotBlank(username)) {
             queryWrapper.like("username", username);
         }
-        return userService.list(queryWrapper);
+        List<User> userList = userService.list(queryWrapper);
+        return userList.stream().map(user -> userService.getSafeUser(user)).collect(Collectors.toList());
     }
 
     @PostMapping("/delete")
-    public boolean deleteUser(@RequestBody long id) { //RequestBody告诉springboot框架，参数绑定到请求体的内容 因为是post
+    public boolean deleteUser(@RequestBody long id, HttpServletRequest request) { //RequestBody告诉springboot框架，参数绑定到请求体的内容 因为是post
+        // 鉴权 仅管理员可查询
+        if (!isAdmin(request)) {
+            return false;
+        }
         if (id < 0) {
             return false;
         }
         return userService.removeById(id);
+    }
+
+    /**
+     * 是否为管理员
+     * @param request request
+     * @return admin or not
+     */
+    private boolean isAdmin(HttpServletRequest request) {
+        Object userObj = request.getSession().getAttribute(USER_LOGIN_STATE);
+        User user = (User) userObj;
+        // user == null 防止空指针异常 NullPointerException
+        return user != null && user.getUserRole() == ADMIN_ROLE;
     }
 }
